#849 Moved GitRepository into namespace SEE.VCS.
