+++ Assets/SEE/Game/City/GitPoller.cs
@@ -6,6 +6,7 @@ using SEE.DataModel.DG;
 using SEE.GO;
 using SEE.GraphProviders;
 using SEE.UI.Notification;
+using SEE.VCS;
 using UnityEngine;
 
 namespace SEE.Game.City
