+++ Assets/SEE/GraphProviders/AllGitBranchesSingleGraphProvider.cs
@@ -9,6 +9,7 @@ using SEE.GraphProviders.VCS;
 using SEE.UI.RuntimeConfigMenu;
 using SEE.Utils;
 using SEE.Utils.Config;
+using SEE.VCS;
 using Sirenix.OdinInspector;
 using Sirenix.Serialization;
 using Sirenix.Utilities;
