+++ Assets/SEE/GraphProviders/Evolution/GitEvolutionGraphProvider.cs
@@ -11,6 +11,7 @@ using SEE.GraphProviders.VCS;
 using SEE.UI.RuntimeConfigMenu;
 using SEE.Utils;
 using SEE.Utils.Config;
+using SEE.VCS;
 using Sirenix.OdinInspector;
 using Sirenix.Serialization;
 using UnityEngine;
