+++ Assets/SEETests/TestGitGraphProvider.cs
@@ -6,6 +6,7 @@ using SEE.Game.City;
 using SEE.GraphProviders.Evolution;
 using SEE.Utils;
 using SEE.Utils.Paths;
+using SEE.VCS;
 using System;
 using System.Collections;
 using System.Collections.Generic;
